# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VIT - Value Intelligence Trust Prediction Engine API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: predictions
    description: Match prediction operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: List all predictions
      description: Returns all past match predictions ordered by recency
      responses:
        "200":
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
    post:
      operationId: createPrediction
      tags: [predictions]
      summary: Create a new prediction
      description: Submit a match for VIT analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePredictionBody"
      responses:
        "201":
          description: Prediction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /predictions/{id}:
    get:
      operationId: getPrediction
      tags: [predictions]
      summary: Get a prediction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prediction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePrediction
      tags: [predictions]
      summary: Delete a prediction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Prediction deleted
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreatePredictionBody:
      type: object
      properties:
        sport:
          type: string
          enum: [football, basketball, tennis]
        homeTeam:
          type: string
        awayTeam:
          type: string
        league:
          type: ["string", "null"]
        matchDate:
          type: ["string", "null"]
      required:
        - sport
        - homeTeam
        - awayTeam
    ScoreSimulation:
      type: object
      properties:
        home:
          type: integer
        away:
          type: integer
        probability:
          type: number
      required:
        - home
        - away
        - probability
    MarketPrediction:
      type: object
      properties:
        market:
          type: string
        prediction:
          type: string
        confidence:
          type: number
        ev:
          type: number
        recommendedOdds:
          type: number
      required:
        - market
        - prediction
        - confidence
        - ev
        - recommendedOdds
    Prediction:
      type: object
      properties:
        id:
          type: integer
        sport:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        league:
          type: ["string", "null"]
        matchDate:
          type: ["string", "null"]
        homeWinProb:
          type: number
        drawProb:
          type: ["number", "null"]
        awayWinProb:
          type: number
        predictedHomeScore:
          type: ["number", "null"]
        predictedAwayScore:
          type: ["number", "null"]
        bestBet:
          type: string
        bestBetEv:
          type: number
        bestBetConfidence:
          type: number
        valueRating:
          type: string
        aiConsensus:
          type: string
        marketPredictions:
          type: array
          items:
            $ref: "#/components/schemas/MarketPrediction"
        scoreSimulations:
          type: array
          items:
            $ref: "#/components/schemas/ScoreSimulation"
        processingTimeMs:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sport
        - homeTeam
        - awayTeam
        - league
        - matchDate
        - homeWinProb
        - drawProb
        - awayWinProb
        - predictedHomeScore
        - predictedAwayScore
        - bestBet
        - bestBetEv
        - bestBetConfidence
        - valueRating
        - aiConsensus
        - marketPredictions
        - scoreSimulations
        - processingTimeMs
        - createdAt