# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIT - Value Intelligence Trust Prediction Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all past match predictions ordered by recency
 * @summary List all predictions
 */
export const ListPredictionsResponseItem = zod.object({
  id: zod.number(),
  sport: zod.string(),
  homeTeam: zod.string(),
  awayTeam: zod.string(),
  league: zod.string().nullable(),
  matchDate: zod.string().nullable(),
  homeWinProb: zod.number(),
  drawProb: zod.number().nullable(),
  awayWinProb: zod.number(),
  predictedHomeScore: zod.number().nullable(),
  predictedAwayScore: zod.number().nullable(),
  bestBet: zod.string(),
  bestBetEv: zod.number(),
  bestBetConfidence: zod.number(),
  valueRating: zod.string(),
  aiConsensus: zod.string(),
  marketPredictions: zod.array(
    zod.object({
      market: zod.string(),
      prediction: zod.string(),
      confidence: zod.number(),
      ev: zod.number(),
      recommendedOdds: zod.number(),
    }),
  ),
  scoreSimulations: zod.array(
    zod.object({
      home: zod.number(),
      away: zod.number(),
      probability: zod.number(),
    }),
  ),
  processingTimeMs: zod.number().nullable(),
  createdAt: zod.date(),
});
export const ListPredictionsResponse = zod.array(ListPredictionsResponseItem);

/**
 * Submit a match for VIT analysis
 * @summary Create a new prediction
 */
export const CreatePredictionBody = zod.object({
  sport: zod.enum(["football", "basketball", "tennis"]),
  homeTeam: zod.string(),
  awayTeam: zod.string(),
  league: zod.string().nullish(),
  matchDate: zod.string().nullish(),
});

/**
 * @summary Get a prediction by ID
 */
export const GetPredictionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPredictionResponse = zod.object({
  id: zod.number(),
  sport: zod.string(),
  homeTeam: zod.string(),
  awayTeam: zod.string(),
  league: zod.string().nullable(),
  matchDate: zod.string().nullable(),
  homeWinProb: zod.number(),
  drawProb: zod.number().nullable(),
  awayWinProb: zod.number(),
  predictedHomeScore: zod.number().nullable(),
  predictedAwayScore: zod.number().nullable(),
  bestBet: zod.string(),
  bestBetEv: zod.number(),
  bestBetConfidence: zod.number(),
  valueRating: zod.string(),
  aiConsensus: zod.string(),
  marketPredictions: zod.array(
    zod.object({
      market: zod.string(),
      prediction: zod.string(),
      confidence: zod.number(),
      ev: zod.number(),
      recommendedOdds: zod.number(),
    }),
  ),
  scoreSimulations: zod.array(
    zod.object({
      home: zod.number(),
      away: zod.number(),
      probability: zod.number(),
    }),
  ),
  processingTimeMs: zod.number().nullable(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a prediction
 */
export const DeletePredictionParams = zod.object({
  id: zod.coerce.number(),
});