# lib/api-zod/src/generated/types/marketPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIT - Value Intelligence Trust Prediction Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketPrediction {
  market: string;
  prediction: string;
  confidence: number;
  ev: number;
  recommendedOdds: number;
}